requirements.txt
Django
Fabric
GitPython
Jinja2
Markdown
MarkupSafe
django_inlinecss
Pillow
PyJWT
PyYAML
Werkzeug
amqp
anyjson
argparse
arrow
backports-abc
backports.ssl-match-hostname
billiard
celery
certifi
cffi
click
crcmod
cryptography
cssselect
decorator
# django-activity-stream
django-annoying
django-appconf
django-bower
django-celery
django-cors-headers
django-crispy-forms
django-environ
django-extensions
django-extra-fields
django-filter
django-geoposition
django-leaflet
django-hstore
django-inmemorystorage
boto
django-storages-redux
django-jsonfield
-e git+https://github.com/agiliq/Django-parsley.git@c11296dafcea475a6a6ec85b016b7a59b56c2470#egg=django_parsley-master
-e git+https://github.com/rosscdh/django-allauth@feature/subject-in-context#egg=django-allauth
django-pipeline==1.6.6
#manifesto
slimit
django-rest-auth
django-rest-swagger
django-rulez
django-shortuuidfield
-e git+https://github.com/rosscdh/django-templated-email@fd2eabd4afa0da520d5e8ffb3dc07a397a2cfbd3#egg=django_templated_email
djangorestframework
djangorestframework-hstore
djangorestframework-jwt
drf-extensions
easy-thumbnails
django-embed-video
ecdsa
enum34
freezegun
future
futures
gitdb
icalendar
idna
influxdb
ipaddress
jsonfield
kombu
livereload
lxml
mkdocs
mock
mock-django
model-mommy
httpretty
ndg-httpsclient==0.3.0
oauthlib
paramiko
-e git+https://github.com/rosscdh/pinax-eventlog.git@d1018d626db926dee6593bc7f51dd653659367b1#egg=pinax_eventlog
# pinax-stripe
psycopg2
pusher
pyOpenSSL
pyasn1
pycparser
pycrypto
pyquery
python-dateutil
python-openid
python-social-auth
pytz
requests
requests-mock
requests-oauthlib
rollbar==0.13.0
shortuuid
singledispatch
six
smmap
# stripe
termcolor
tornado
urllib3
wsgiref
django-rq
honcho
django-autoslug
py-moneyed
django-money
-e git+https://github.com/rosscdh/django-templated-email.git@fd2eabd4afa0da520d5e8ffb3dc07a397a2cfbd3#egg=templated_email
django-countries
-e git+https://github.com/batiste/django-page-cms.git@fd94430ee8e1877acb285a54060397b8c08bb8a4#egg=pages
django-taggit
html2text
coverage
awesome-slugify==1.6.5
django-versatileimagefield==1.5
# PDF Rendering
django-easy-pdf
xhtml2pdf>=0.0.6
reportlab>=2.7,<3
# End PDF Rendering
-e git+https://github.com/rosscdh/python-secupay.git@master#egg=secupay